Added automation scripts.
[Details]
Added the automation scripts that allow the defects4j multifault dataset
to be processed automatically and the coverage results stored.

[Note]
The automation process does not currently "remove" bugs that are not in
the spectrum from the dataset, and instead just does not identify them
and prints a message to this effect. It is not known whether this option
or removing them is better.

[File's changed]

	new file:   dump_versions.py
	new file:   gen_coverage.sh
	new file:   single_coverage.sh