Updated dump_versions.
[Details]
Updated the dump_versions script to display all the versions for this
dataset. This includes the bugs in each version that have transplanted
test cases AND identification on at least one line. Had to edit the
backtrack script in order to make it callable from dump_versions for
this purpose.

[File's changed]

	modified:   defects4j_multi_with_jars.patch
	modified:   dump_versions.py