Cleaned up dump_versions and d4j patch.
[Details]
Cleaned up the dump_versions script so that it can be reused as a python
module if needed. Also moved all the scripts added to defects4j to a
scripts/ directory inside defects4j/framework/bin. This helps to clean
up the changes to defects4j and restricts what is accessible to the
users environment (since everything in framework/bin is exposed). This
commit also fixes a few issues in the defects4j_multi script, and adds
some functionality for external test suite running.

[File's changed]
	modified:   defects4j_multi_with_jars.patch
	modified:   dump_versions.py