Updated multi checkout to take a version number.
[Details]
The multi checkout command previously could either take a list of bugs,
or a "version" which was just the name of the project concatenated with
the list of bugs. The command is now changed to take either a list of
bugs, or just the version number, from which the list of bugs is
retrieved from the dump_versions script. This makes the script
follow the same format as the bugsinpy script.

[File's changed]
	modified:   defects4j_multi_with_jars.patch